Sponsors are now updated by name, not unique ID
This is a TEMPORARY solution; the CRP unique ID is not actually unique right now for candidate
committees.  Rather, it uses the candidate's ID for the committee unique ID.  This is going
to be changed in the next few weeks on CRPs end, but for now it is causing candidates with
more than one committee (e.g. ted cruz) to have one of their committees missing.